requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
pandas-ta>=0.3.0
numpy>=1.24.0
yfinance>=0.2.0
apscheduler>=3.10.0
kiteconnect>=4.1.0
growwapi>=0.1.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0
pytz>=2023.3
redis>=5.0.0
celery>=5.3.0
diskcache>=5.6.0
google-genai>=1.0.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
lxml>=4.9.0
python-telegram-bot>=20.0

python-jose
contextvars
pyjwt
booktype